requirements.txt
motor~=3.7.1
PyJWT~=2.10.1
pymongo~=4.13.0
fastapi~=0.115.12
passlib~=1.7.4
pydantic~=2.11.5
uvicorn~=0.34.3
joblib~=1.5.1
numpy~=2.3.0
pandas~=2.3.0
pydantic[email]
scikit-learn
uvicorn[standard]
websockets
